test(product-int): update the integration test for products
the integration test for products have they have been broken since the pagination update

partly resolves #5
